+++ Source/ACE.ACClientLauncher/Subscription.cs
@@ -1,11 +1,7 @@
-﻿using Newtonsoft.Json;
+using Newtonsoft.Json;
 using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
-namespace ACE.CmdLineLauncher
+namespace ACE.ACClientLauncher
 {
     public class Subscription
     {