+++ Source/ACE.Database.Tests/ContentTests.cs
@@ -1,4 +1,4 @@
-﻿using ACE.Common;
+using ACE.Common;
 using ACE.Entity;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System;
@@ -19,7 +19,7 @@ namespace ACE.Database.Tests
         public static void TestSetup(TestContext context)
         {
             // copy config.json
-            File.Copy(Path.Combine(Environment.CurrentDirectory, "..\\..\\..\\..\\ACE\\Config.json"), ".\\Config.json", true);
+            File.Copy(Path.Combine(Environment.CurrentDirectory, "..\\..\\..\\..\\..\\ACE\\Config.json"), ".\\Config.json", true);
 
             ConfigManager.Initialize();
             worldDb = new WorldDatabase();