+++ Source/ACE.Database/Database.cs
@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using System.Collections.Generic;
 using System.Data;
 using System.Diagnostics;
@@ -12,13 +12,14 @@ using ACE.Common;
 using ACE.Common.Extensions;
 using System.Reflection;
 using System.Data.SqlClient;
+using ACE.Database.Extensions;
 
 namespace ACE.Database
 {
     public class Database
     {
         // This is a debug channel for the general debugging of the database.
-        private ILog log = LogManager.GetLogger("Database");
+        private ILog log = LogManager.GetLogger(System.Reflection.MethodBase.GetCurrentMethod().DeclaringType);
 
         private static readonly Dictionary<Type, List<Tuple<PropertyInfo, DbFieldAttribute>>> propertyCache = new Dictionary<Type, List<Tuple<PropertyInfo, DbFieldAttribute>>>();
 
@@ -977,7 +978,7 @@ namespace ACE.Database
             }
         }
 
-        protected MySqlResult SelectPreparedStatement<T>(T id, params object[] parameters)
+        protected DataTable SelectPreparedStatement<T>(T id, params object[] parameters)
         {
             // Debug.Assert(typeof(T) == PreparedStatementType, "Invalid prepared statement type.");
 
@@ -1000,10 +1001,9 @@ namespace ACE.Database
 
                         using (var commandReader = command.ExecuteReader(CommandBehavior.Default))
                         {
-                            using (var result = new MySqlResult())
+                            using (var result = new DataTable())
                             {
-                                result.Load(commandReader);
-                                result.Count = (uint)result.Rows.Count;
+                                result.LoadEx(commandReader);
                                 return result;
                             }
                         }
@@ -1019,7 +1019,7 @@ namespace ACE.Database
             return null;
         }
 
-        protected async Task<MySqlResult> SelectPreparedStatementAsync<T>(T id, params object[] parameters)
+        protected async Task<DataTable> SelectPreparedStatementAsync<T>(T id, params object[] parameters)
         {
             // Debug.Assert(typeof(T) == PreparedStatementType, "Invalid prepared statement type.");
 
@@ -1044,10 +1044,9 @@ namespace ACE.Database
                         {
                             using (var commandReader = command.ExecuteReader(CommandBehavior.Default))
                             {
-                                using (var result = new MySqlResult())
+                                using (var result = new DataTable())
                                 {
-                                    result.Load(commandReader);
-                                    result.Count = (uint)result.Rows.Count;
+                                    result.LoadEx(commandReader);
                                     return result;
                                 }
                             }